+++ standman/listener.py
@@ -1,0 +1,8 @@
+
+from typing import Protocol
+
+
+class Listener(Protocol):
+    def listen(self, tag: str, *args, **kwargs) -> None:
+        ...
+