+++ standman/process_observer/__init__.py
@@ -1,0 +1,9 @@
+
+from .observer import ProcessObserver
+from .observer import create_process_observer
+
+__all__ = (
+    'ProcessObserver',
+    'create_process_observer'
+)
+